+++ packages/coding-agent/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `providers.webSearchFanout` to query the first N eligible web search providers concurrently with separate results and graceful partial failures ([#11087](https://github.com/can1357/oh-my-pi/pull/11087) by [@mustafaabidali](https://github.com/mustafaabidali)).
+
 ## [18.1.12] - 2026-09-06
 
 - Fixed edit and write results to report the formatted bytes actually committed by LSP writethrough.